+++ src/Scenes/LevelSelect/LevelCard.h
@@ -5,7 +5,7 @@
 class LevelCard : public Aegis::Widget
 {
 public:
-	LevelCard(int level, const std::array<double, 2>& star_thresholds, const double fastest_time_);
+	LevelCard(int level, const std::array<int, 2>& star_thresholds, const double fastest_time_);
 
 	void Render() const override;
 	void OnEvent(Aegis::Event& event) override;
